+++ src/main/java/com/github/wintersteve25/tau/renderer/HudUIRenderer.java
@@ -1,5 +1,6 @@
 package com.github.wintersteve25.tau.renderer;
 
+import com.github.wintersteve25.tau.theme.ColorScheme;
 import com.github.wintersteve25.tau.theme.DefaultColorScheme;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.MainWindow;
@@ -16,19 +17,25 @@ public class HudUIRenderer {
     private final UIComponent uiComponent;
     private final List<IRenderable> components;
     private final List<DynamicUIComponent> dynamicUIComponents;
+    private final ColorScheme colorScheme;
 
     private boolean built;
     private int screenWidth;
     private int screenHeight;
 
-    public HudUIRenderer(UIComponent uiComponent) {
+    public HudUIRenderer(UIComponent uiComponent, ColorScheme colorScheme) {
         this.uiComponent = uiComponent;
         this.components = new ArrayList<>();
         this.dynamicUIComponents = new ArrayList<>();
+        this.colorScheme = colorScheme;
+    }
+    
+    public HudUIRenderer(UIComponent uiComponent) {
+        this(uiComponent, DefaultColorScheme.INSTANCE);
     }
 
     private void init() {
-        Layout layout = new Layout(screenWidth, screenHeight, DefaultColorScheme.INSTANCE);
+        Layout layout = new Layout(screenWidth, screenHeight, colorScheme);
 
         components.clear();
         dynamicUIComponents.clear();