+++ src/main/java/com/github/wintersteve25/tau/theme/DefaultColorScheme.java
@@ -8,7 +8,9 @@ public class DefaultColorScheme implements ColorScheme {
     public static final DefaultColorScheme INSTANCE = new DefaultColorScheme();
     
     private static final Color BACKGROUND = new Color(0xFF252525);
-    private static final Color SECONDARY_BACKGROUND = new Color(0xFF373737);
+    private static final Color ON_BACKGROUND = new Color(0xFF373737);
+    private static final Color SECONDARY_BACKGROUND = new Color(0xFF424242);
+    private static final Color ON_SECONDARY_BACKGROUND = new Color(0xFF565656);
     private static final Color TEXT = new Color(0xFFE8E8E8);
     private static final Color INTERACTABLE = new Color(0xFF8B8B8B); 
     private static final Color HOVER = new Color(0xFFC6C6C6); 
@@ -25,6 +27,16 @@ public class DefaultColorScheme implements ColorScheme {
     @Override
     public Color secondaryBackgroundColor() {
         return SECONDARY_BACKGROUND;
+    }
+
+    @Override
+    public Color onBackground() {
+        return ON_BACKGROUND;
+    }
+
+    @Override
+    public Color onSecondary() {
+        return ON_SECONDARY_BACKGROUND;
     }
 
     @Override